requirements.txt
asgiref==3.11.1
celery==5.4.0
Django==6.0.2
djangorestframework==3.16.1
pillow==12.1.1
psycopg2-binary==2.9.11
redis==5.2.1
sqlparse==0.5.5
